+++ bit_vector.h
@@ -1,6 +1,7 @@
 #ifndef GAUCHE_BIT_VECTOR_H
 #define GAUCHE_BIT_VECTOR_H
 
+#include <sys/types.h>
 #include <gauche.h>
 #include <gauche/extend.h>
 
@@ -8,11 +9,11 @@ SCM_DECL_BEGIN
 
 typedef struct ScmBitVectorRec {
   SCM_HEADER;
-  int length;
-  unsigned int builded : 1;
-  int numBlocks : (SIZEOF_INT*CHAR_BIT-1);
+  long length;
+  u_long builded : 1;
+  long numBlocks : (SIZEOF_INT*CHAR_BIT-1);
   ScmBits *blocks;
-  int *rankTable;
+  long *rankTable;
 } ScmBitVector;
 
 SCM_CLASS_DECL(Scm_BitVectorClass);
@@ -25,16 +26,16 @@ SCM_CLASS_DECL(Scm_BitVectorClass);
 #define SCM_BIT_VECTOR_BLOCKS(obj)     (SCM_BIT_VECTOR(obj)->blocks)
 #define SCM_BIT_VECTOR_RANK_TABLE(obj) (SCM_BIT_VECTOR(obj)->rankTable)
 
-SCM_EXTERN ScmBitVector *Scm_MakeBitVector(int length, int fill);
+SCM_EXTERN ScmBitVector *Scm_MakeBitVector(long length, int fill);
 SCM_EXTERN ScmBitVector *Scm_BitVectorClear(ScmBitVector *bv);
 SCM_EXTERN ScmBitVector *Scm_BitVectorBuild(ScmBitVector *bv);
-SCM_EXTERN int Scm_BitVectorRef(ScmBitVector *bv, int i);
-SCM_EXTERN ScmBitVector *Scm_BitVectorSet(ScmBitVector *bv, int i, int bit);
+SCM_EXTERN int Scm_BitVectorRef(ScmBitVector *bv, long i);
+SCM_EXTERN ScmBitVector *Scm_BitVectorSet(ScmBitVector *bv, long i, int bit);
 SCM_EXTERN ScmBitVector *Scm_BitVectorFill(ScmBitVector *bv,
-                                           int fill, int start, int end);
-SCM_EXTERN int Scm_BitVectorRank(ScmBitVector *bv, int i, int bit);
-SCM_EXTERN int Scm_BitVectorRankAll(ScmBitVector *bv, int bit);
-SCM_EXTERN int Scm_BitVectorSelect(ScmBitVector *bv, int i, int bit);
+                                           int fill, long start, long end);
+SCM_EXTERN long Scm_BitVectorRank(ScmBitVector *bv, long i, int bit);
+SCM_EXTERN long Scm_BitVectorRankAll(ScmBitVector *bv, int bit);
+SCM_EXTERN long Scm_BitVectorSelect(ScmBitVector *bv, long i, int bit);
 
 SCM_DECL_END
